change Popup file to index && css file
